Version 0.0.2 a set-point before the next refactor
It's more important that the generated odp files are compatible with
Excel LibreOffice and OO than it is to have each page of each
presentation be fully custimizable.

I'm starting a refactor where every page in the presentation will have
the same background (solid, gradient or bitmap) but all 3 apps will open
them pretty much the same.